+++ changelog.md
@@ -10,6 +10,13 @@
   as expected, the container path was not made absolute so starting
   the container failed.
 
+- Fixed passing of the `PS1` environment variable (`--env
+  PS1=...`). The entrypoint script has `#!/bin/sh` which means the
+  script is not executed as interactive, and then PS1 is not part of
+  the environment in the script, and hence not in the command (say
+  bash, for interactive use). Now we handle this as a special case.
+
+
 ### Changed
 
 - Moved `--verbose` and `--quiet` to be global options. Now they need